+++ lib/opal/parser.rb
@@ -327,36 +327,78 @@ module Opal
       end
     end
 
-    def new_args(norm, opt, rest, block)
+    def new_args_tail(kwarg, kwrest, block)
+      [kwarg, kwrest, block]
+    end
+
+    def new_args(norm, opt, rest, tail)
       res = s(:args)
 
       if norm
         norm.each do |arg|
           scope.add_local arg
-          res << arg
+          res << s(:arg, arg)
         end
       end
 
       if opt
         opt[1..-1].each do |_opt|
-          res << _opt[1]
+          res << s(:optarg, _opt[1], _opt[2])
         end
       end
 
       if rest
-        res << rest
-        rest_str = rest.to_s[1..-1]
-        scope.add_local rest_str.to_sym unless rest_str.empty?
+        restname = rest.to_s[1..-1]
+
+        if restname.empty?
+          res << s(:restarg)
+        else
+          res << s(:restarg, restname.to_sym)
+          scope.add_local restname.to_sym
+        end
       end
 
-      if block
-        res << block
-        scope.add_local block.to_s[1..-1].to_sym
+      # kwarg
+      if tail and tail[0]
+        tail[0].each do |kwarg|
+          res << kwarg
+        end
       end
 
-      res << opt if opt
+      # kwrestarg
+      if tail and tail[1]
+        res << tail[1]
+      end
+
+      # block
+      if tail and tail[2]
+        blockname = tail[2].to_s[1..-1].to_sym
+        scope.add_local blockname
+        res << s(:blockarg, blockname)
+      end
 
       res
+    end
+
+    def new_kwarg(name)
+      scope.add_local name[1]
+      s(:kwarg, name[1])
+    end
+
+    def new_kwoptarg(name, val)
+      scope.add_local name[1]
+      s(:kwoptarg, name[1], val)
+    end
+
+    def new_kwrestarg(name = nil)
+      result = s(:kwrestarg)
+
+      if name
+        scope.add_local name[0].to_sym
+        result << name[0].to_sym
+      end
+
+      result
     end
 
     def new_block_args(norm, opt, rest, block)